refactor(apps): share sprite renderer
