+++ website/src/apps/crew-companion/GalleryPanel.tsx
@@ -21,7 +21,7 @@ import { applySvgColorMap, type ColorMap } from './colorCustomizer'
 import { ghostPoseForKey } from './ghostEyes'
 import { GhostEyeOverlay } from './GhostEyeOverlay'
 import { LottieRenderer } from './LottieRenderer'
-import { SpriteRenderer } from './SpriteRenderer'
+import { SpriteRenderer } from '../shared/SpriteRenderer'
 import { PackEditor } from './PackEditor'
 import { SpriteImporter } from './SpriteImporter'
 import { buildSpritePackData, firstFramePreview } from './petdexImport'
