+++ website/src/apps/crew-companion/PetAvatar.tsx
@@ -22,7 +22,7 @@ import { HIDES_EYES, type GhostAccessory } from './ghostAccessories'
 import { ghostPoseForKey, ghostEyeOffsetFor, POSED_ANIMS } from './ghostEyes'
 import { animClassFor, type PetAnim } from './petAnim'
 import { LottieRenderer } from './LottieRenderer'
-import { SpriteRenderer } from './SpriteRenderer'
+import { SpriteRenderer } from '../shared/SpriteRenderer'
 import { toDataUri } from './animationResolver'
 import { applySvgColorMap, type ColorMap } from './colorCustomizer'
 import './petMotion.css'
