+++ website/src/apps/mochi/src/renderer/PetWidget.tsx
@@ -33,7 +33,7 @@ import {
   resolvePetName,
 } from '../../builtinPacks'
 import { LottieRenderer } from './LottieRenderer'
-import { SpriteRenderer } from './SpriteRenderer'
+import { SpriteRenderer } from '../../../shared/SpriteRenderer'
 import { PetContextMenu } from './PetContextMenu'
 
 // ── Hardcoded SVG fallbacks for default-mochi built-in pack ────────────────
