+++ website/src/apps/mochi/src/renderer/SpriteImporter.tsx
@@ -4,7 +4,7 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react'
 import { FolderOpen, Plus} from 'lucide-react'
 import { REQUIRED_STATES, ALL_MOODS, OPTIONAL_STATES, type PackMeta } from '../shared/appearanceTypes'
-import { SpriteRenderer } from './SpriteRenderer'
+import { SpriteRenderer } from '../../../shared/SpriteRenderer'
 import { PackInfoHeader } from './PackInfoHeader'
 import { SaveDialog } from './SaveDialog'
 import { EditorFooter } from './EditorFooter'
